coach-engine/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
sqlalchemy==2.0.28
python-chess==1.999
reportlab==4.1.0
pydantic==2.6.4
python-multipart==0.0.9
PyJWT==2.8.0
httpx==0.27.0
python-dotenv==1.0.1
